# api/openapi/http-api.yaml
openapi: 3.0.2
servers:
  - url: /v1
info:
  version: 1.0.0
  title: OpenEverest schema
  description: |
    # Authentication
    All requests to Everest API require `Authorization: Bearer <token>` header with a valid token in plain-text.
    
    The token can be obtained by using `everestctl token reset` which resets the token and prints it to the screen.
...
security:
  - BearerAuth: []
paths:
...
  '/clusters/{cluster}/namespaces/{namespace}/instances':
    x-everest-resource-name: instances
    get:
      tags:
        - Instance
      summary: List instances
      description: |
        This API lists all database instances in the specified namespace and cluster.
      operationId: listInstances
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: namespace
          in: path
          required: true
          description: The namespace where instances are located
          schema:
            type: string
      responses:
        '200':
          description: List of instances
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/InstanceList'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Instance
      summary: Create instance
      description: |
        This API creates a new database instance in the specified namespace and cluster.
      operationId: createInstance
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: namespace
          in: path
          required: true
          description: The namespace where the instance will be created
          schema:
            type: string
      requestBody:
        description: The instance object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: 'crds.gen.yaml#/components/schemas/Instance'
      responses:
        '200':
          description: Instance created successfully
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/Instance'
        '201':
          description: Instance created successfully
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/Instance'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/clusters/{cluster}/namespaces/{namespace}/instances/{instance}':
    x-everest-resource-name: instances
    get:
      tags:
        - Instance
      summary: Get instance
      description: |
        This API gets the database instance specified by the `instance` name in the specified `namespace` and `cluster`.
      operationId: getInstance
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: namespace
          in: path
          required: true
          description: The namespace where the instance is located
          schema:
            type: string
        - name: instance
          in: path
          required: true
          description: The name of the instance
          schema:
            type: string
      responses:
        '200':
          description: Instance details
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/Instance'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Instance
      summary: Update instance
      description: |
        This API updates the database instance specified by the `instance` name in the specified `namespace` and `cluster`, replacing the entire resource with the new one.
        To change only specific fields, use PATCH instead.
      operationId: updateInstance
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: namespace
          in: path
          required: true
          description: The namespace where the instance is located
          schema:
            type: string
        - name: instance
          in: path
          required: true
          description: The name of the instance
          schema:
            type: string
      requestBody:
        description: The instance object to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: 'crds.gen.yaml#/components/schemas/Instance'
      responses:
        '200':
          description: Instance updated successfully
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/Instance'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
        - Instance
      summary: Patch instance
      description: |
        This API applies Merge Patch to the database instance specified by the `instance` name in the specified `namespace` and `cluster`.
        Members absent from the patch keep their current value, and a member set to `null` is removed. A path that does not exist on the Instance is
        rejected rather than ignored. Include `metadata.resourceVersion` to make the patch conditional, which fails with 409 if the instance has changed.
      operationId: patchInstance
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: namespace
          in: path
          required: true
          description: The namespace where the instance is located
          schema:
            type: string
        - name: instance
          in: path
          required: true
          description: The name of the instance
          schema:
            type: string
      requestBody:
        description: |
          A partial Instance document. `status` and the `ownerReferences`, `finalizers`, `name` and `namespace` members of `metadata` are rejected.
        required: true
        content:
          application/merge-patch+json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Instance patched successfully
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/Instance'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '415':
          description: The Content-Type is not application/merge-patch+json
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Instance
      summary: Delete instance
      description: |
        This API deletes the database instance specified by the `instance` name in the specified `namespace` and `cluster`.
      operationId: deleteInstance
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: namespace
          in: path
          required: true
          description: The namespace where the instance is located
          schema:
            type: string
        - name: instance
          in: path
          required: true
          description: The name of the instance
          schema:
            type: string
        - name: deletionPolicy
          in: query
          required: false
          description: |
            Override the instance's spec.deletionPolicy before deletion.
            "Cascade" (default) deletes all Backup and Restore CRs referencing this instance.
            "Orphan" leaves Backup and Restore CRs in place.
            When omitted, the existing value on the Instance CR is used.
          schema:
            type: string
            enum:
              - Cascade
              - Orphan
      responses:
        '200':
          description: Instance deleted successfully
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/Instance'
        '204':
          description: Instance deleted successfully
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/clusters/{cluster}/namespaces/{namespace}/instances/{instance}/connection':
    x-everest-resource-name: instances
    get:
      tags:
        - Instance
      summary: Get instance connection details
      description: |
        This API returns the connection details for the database instance specified by the `instance` name
        in the specified `namespace` and `cluster`. The connection details include the host, port, username,
        password, and a pre-built connection URI. These details are populated by the provider once the instance
        is running and ready.
      operationId: getInstanceConnection
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: namespace
          in: path
          required: true
          description: The namespace where the instance is located
          schema:
            type: string
        - name: instance
          in: path
          required: true
          description: The name of the instance
          schema:
            type: string
      responses:
        '200':
          description: Connection details for the instance
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/InstanceConnectionDetails'
        '404':
          description: Instance or connection details not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      description: Error response
      properties:
        message:
          type: string
...